OK-609 Перевести easy-redux на typescript
