+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "mocha -r ts-node/register test/**/*",
     "build": "babel src --out-dir lib",
-    "dev": "better-npm-run dev"
+    "dev": "better-npm-run dev",
+    "lint": "tslint --project tsconfig.json"
   },
   "betterScripts": {
     "dev": {
@@ -34,16 +35,18 @@
   },
   "homepage": "https://github.com/welljs/easy-redux#readme",
   "devDependencies": {
-    "@types/chai": "^4.0.4",
-    "@types/mocha": "^2.2.43",
-    "@types/redux": "^3.6.0",
-    "chai": "^4.1.2",
-    "mocha": "^3.5.3",
-    "ts-lint": "^4.5.1",
-    "ts-node": "^3.3.0",
-    "typescript": "^2.5.2"
+    "@types/chai": "4.1.5",
+    "@types/mocha": "5.2.5",
+    "@types/redux": "3.6.31",
+    "chai": "4.1.2",
+    "mocha": "5.2.0",
+    "tslint": "5.11.0",
+    "tslint-loader": "3.6.0",
+    "tslint-react": "3.6.0",
+    "ts-node": "7.0.1",
+    "typescript": "3.0.3"
   },
   "dependencies": {
-    "redux": "^3.5.2"
+    "redux": "4.0.0"
   }
 }
