+++ src/index.ts
@@ -1,18 +1,4 @@
-export const isFunction = (thing: any): boolean => typeof thing === 'function';
-
-export {default as middleware} from './middleware';
-
-export {
-  applyReducer,
-  localCompose,
-  combine,
-  mergeReducer
-} from './easy-redux';
-
-export {createAction, createActions} from './actionCreator';
-
-export {
-  applyRequestReducers,
-  createRequestReducer,
-  requestReducer
-} from './requestReducer';
+export * from './middleware';
+export * from './easy-redux';
+export * from './actionCreator';
+export * from './requestReducer';
