+++ tsconfig.json
@@ -1,16 +1,18 @@
 {
   "compilerOptions": {
-    "outDir": "./dist/",
+    "allowSyntheticDefaultImports": false,
     "module": "es6",
     "target": "es5",
     "lib":["dom", "es6", "dom.iterable", "scripthost", "es2016.array.include", "es2015.promise", "es2017"],
-    "sourceMap": false,
+    "sourceMap": true,
     "strictNullChecks": true,
     "jsx": "react",
-    "allowJs": true,
     "moduleResolution": "node",
     "experimentalDecorators": true,
-    "removeComments": true
+    "removeComments": true,
+    "outDir": "./_tscache/",
+    "declaration": true,
+    "declarationDir": "./lib"
   },
   "include": [
     "./src",
